Completed implementation of Pasqal Cryo
The implementation of the Pasqal Cryo in SilvaBot is completed. With the GUI, the 8 temperatures available from the cryo can be seen in the cryostat device drop down (if 999 is displayed, it means that the Pasqal cryo sent a nan temperature value). In the device drop down menu of the cryostat, the target temperature of loop one can be changed by modifying the value of Set_T. In addition, the compressor can be turned on or off by changing the value of OnOff_comp (On = 1, Off = 0). The same can be done with the heaters that follow loop 1.