Finished the PI863 implementation
The PI863 translation stage can be controlled in the following ways. Changing the movement speed of the stage, moving the stage to a given position, making the stage do a scan (default back an forth over the whole length of the stage - can be change in the futur). In addition, the position of the stage is displayed on the GUI.